requirements.txt
elasticsearch==7.13.1
elasticsearch-dsl==7.3.0
singer-python==5.9.1
